requirements_fastapi.txt
# FastAPI Migration Dependencies
# Install with: pip install -r requirements_fastapi.txt

# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Template and Static Files
jinja2==3.1.3
aiofiles==23.2.1

# Rate Limiting
slowapi==0.1.9

# Session Management
itsdangerous>=2.1.2
starlette-session==0.4.3

# CSRF Protection


# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Async Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0

# Flask-SQLAlchemy (for hybrid migration - still using models from Flask app)
Flask-SQLAlchemy==3.1.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-Limiter==3.12.0
WTForms==3.1.1

# Optional: Redis for session storage
redis==5.0.1
hiredis==2.3.2

# HTTP Client for async operations
aiohttp==3.9.1

# Image Processing
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.32.4
bcrypt==4.0.1
bleach==6.1.0
markdown==3.5.1
vivosun_thermo>=1.0.0

# Testing
pytest-asyncio==0.23.3
httpx==0.26.0
pytest==8.0.0
pytest-cov==4.1.0

# Code Quality and Linting
flake8==7.0.0
pylint==3.0.3
black==24.2.0
isort==5.13.2
mypy==1.8.0

# Security Analysis
bandit==1.7.6
safety==2.3.5

# Database Migrations
alembic==1.13.1
psycopg2-binary==2.9.9
